+++ src/library/DIPS.Mobile.UI/Components/BottomSheets/BottomSheetNavigationEntry.cs
@@ -1,0 +1,3 @@
+namespace DIPS.Mobile.UI.Components.BottomSheets;
+
+internal record BottomSheetNavigationEntry(ContentPage Page);